requirements.txt
coloredlogs
fuzzywuzzy
google-cloud-aiplatform
openai
pandas
pdoc3
python-Levenshtein
pyyaml
streamlit
